+++ kv/cursor.go
@@ -13,8 +13,6 @@ type staticCursor struct {
 }
 
 // Pair is a struct for key value pairs.
-// It also includes an error which should only be non-nil
-// when the Key and Value are nil.
 type Pair struct {
 	Key   []byte
 	Value []byte